Update Scene_01_DesertStart.unity
any changes from now on to be made on this version.
the walking animations are fixed now edit the rest of the player animations fixed the jump animation.
then move towards the game art, set the background add, terrain objects, then set the  background scroll with respective to player.
then focus on the movement of the character hoe it walks on the ground. For testing purpose I had added a temporary ground for Nirwar to stand on. Add a appropriate tilemap  idk how you guys handle the map thing but its on you  guys so figure accordingly.
while handling the game art remember to follow hierarchy.
For world A all the objects go to Environment present.
For World B all the Objects go to environment Past.
when you guys add the background do keep this in mind.
the layering is as follows.
 player is layer 10.
 ground is 0 for world A objects.
 ground if -1 for world B objects.
further more do cross check all of this from inspector for both the environments.